Move tests to directories.
