+++ test/bitcorenode/server.js
@@ -4,13 +4,12 @@ var should = require('chai').should();
 var proxyquire = require('proxyquire');
 var bitcore = require('bitcore');
 var sinon = require('sinon');
-var Service = require('../bitcorenode');
-var Server = require('../bitcorenode/server.js');
+var Server = require('../../bitcorenode/server.js');
 
 describe('Server Utilities', function() {
 
   describe('#readHttpsOptions', function() {
-    var server = proxyquire('../bitcorenode/server.js', {
+    var server = proxyquire('../../bitcorenode/server.js', {
       fs: {
         readFileSync: function(arg) {
           return arg;
@@ -38,7 +37,7 @@ describe('Server Utilities', function() {
   });
   describe('#startCluster', function() {
     it('will throw error with missing options', function() {
-      var server = proxyquire('../bitcorenode/server.js', {});
+      var server = proxyquire('../../bitcorenode/server.js', {});
       (function() {
         server.startCluster({});
       }).should.throw('When running in cluster mode, locker server');
@@ -67,7 +66,7 @@ describe('Server Utilities', function() {
       var httpServer = {
         listen: sinon.stub()
       };
-      var server = proxyquire('../bitcorenode/server.js', {
+      var server = proxyquire('../../bitcorenode/server.js', {
         'os': {
           cpus: sinon.stub().returns({length: 8})
         },
@@ -102,7 +101,7 @@ describe('Server Utilities', function() {
       function TestWSApp() {}
       TestWSApp.prototype.start = sinon.stub().callsArg(2);
       var listen = sinon.stub().callsArg(1);
-      var server = proxyquire('../bitcorenode/server.js', {
+      var server = proxyquire('../../bitcorenode/server.js', {
         '../lib/expressapp': TestExpressApp,
         '../lib/wsapp': TestWSApp,
         'http': {
@@ -135,7 +134,7 @@ describe('Server Utilities', function() {
       function TestWSApp() {}
       TestWSApp.prototype.start = sinon.stub().callsArg(2);
       var listen = sinon.stub().callsArg(1);
-      var server = proxyquire('../bitcorenode/server.js', {
+      var server = proxyquire('../../bitcorenode/server.js', {
         '../lib/expressapp': TestExpressApp,
         '../lib/wsapp': TestWSApp,
         'http': {
@@ -158,7 +157,7 @@ describe('Server Utilities', function() {
       function TestWSApp() {}
       TestWSApp.prototype.start = sinon.stub().callsArgWith(2, new Error('test'));
       var listen = sinon.stub().callsArg(1);
-      var server = proxyquire('../bitcorenode/server.js', {
+      var server = proxyquire('../../bitcorenode/server.js', {
         '../lib/expressapp': TestExpressApp,
         '../lib/wsapp': TestWSApp,
         'http': {
@@ -192,7 +191,7 @@ describe('Server Utilities', function() {
           listen: listen
         };
       };
-      var server = proxyquire('../bitcorenode/server.js', {
+      var server = proxyquire('../../bitcorenode/server.js', {
         '../lib/expressapp': TestExpressApp,
         '../lib/wsapp': TestWSApp,
         'https': {
@@ -211,159 +210,4 @@ describe('Server Utilities', function() {
     });
   });
 
-});
-
-
-describe('Bitcore Node Service', function() {
-  describe('#constructor', function() {
-    it('https settings from node', function() {
-      var node = {
-        https: true,
-        httpsOptions: {
-          key: 'key',
-          cert: 'cert'
-        }
-      };
-      var options = {
-        node: node
-      };
-      var service = new Service(options);
-      service.node.should.equal(node);
-      service.options.should.equal(options);
-    });
-  });
-  describe('#_getConfiguration', function() {
-    it('will throw with an unknown network', function() {
-      var options = {
-        node: {
-          network: 'unknown'
-        }
-      };
-      var service = new Service(options);
-      (function() {
-        service._getConfiguration();
-      }).should.throw('Unknown network');
-    });
-    it('livenet local insight', function() {
-      var options = {
-        node: {
-          network: bitcore.Networks.livenet,
-          port: 3001
-        }
-      };
-      var service = new Service(options);
-      var config = service._getConfiguration();
-      config.blockchainExplorerOpts.livenet.should.deep.equal({
-        'apiPrefix': '/insight-api',
-        'provider': 'insight',
-        'url': 'http://localhost:3001'
-      });
-    });
-    it('testnet local insight', function() {
-      var options = {
-        node: {
-          network: bitcore.Networks.testnet,
-          port: 3001
-        }
-      };
-      var service = new Service(options);
-      var config = service._getConfiguration();
-      config.blockchainExplorerOpts.testnet.should.deep.equal({
-        'apiPrefix': '/insight-api',
-        'provider': 'insight',
-        'url': 'http://localhost:3001'
-      });
-    });
-  });
-  describe('#start', function(done) {
-    it('error from configuration', function(done) {
-      var options = {
-        node: {}
-      };
-      var service = new Service(options);
-      service._getConfiguration = function() {
-        throw new Error('test');
-      };
-      service.start(function(err) {
-        err.message.should.equal('test');
-        done();
-      });
-    });
-    it('error from blockchain monitor', function(done) {
-      var app = {};
-      function TestBlockchainMonitor() {}
-      TestBlockchainMonitor.prototype.start = sinon.stub().callsArgWith(1, new Error('test'));
-      function TestLocker() {}
-      TestLocker.prototype.listen = sinon.stub();
-      function TestEmailService() {}
-      TestEmailService.prototype.start = sinon.stub();
-      var TestService = proxyquire('../bitcorenode', {
-        '../lib/blockchainmonitor': TestBlockchainMonitor,
-        '../lib/emailservice': TestEmailService,
-        'socket.io': sinon.stub().returns({
-          on: sinon.stub()
-        }),
-        'locker-server': TestLocker,
-      });
-      var options = {
-        node: {}
-      };
-      var service = new TestService(options);
-      var config = {
-        cluster: true
-      };
-      service._getConfiguration = sinon.stub().returns(config);
-      service._startWalletService = sinon.stub().callsArg(1);
-      service.start(function(err) {
-        err.message.should.equal('test');
-        done();
-      });
-    });
-    it('error from email service', function(done) {
-      var app = {};
-      function TestBlockchainMonitor() {}
-      TestBlockchainMonitor.prototype.start = sinon.stub().callsArg(1);
-      function TestLocker() {}
-      TestLocker.prototype.listen = sinon.stub();
-      function TestEmailService() {}
-      TestEmailService.prototype.start = sinon.stub().callsArgWith(1, new Error('test'));
-      var TestService = proxyquire('../bitcorenode', {
-        '../lib/blockchainmonitor': TestBlockchainMonitor,
-        '../lib/emailservice': TestEmailService,
-        'socket.io': sinon.stub().returns({
-          on: sinon.stub()
-        }),
-        'locker-server': TestLocker,
-      });
-      var options = {
-        node: {},
-      };
-      var service = new TestService(options);
-      service._getConfiguration = sinon.stub().returns({
-        emailOpts: {},
-        lockerOpts: {
-          lockerServer: {
-            port: 3231
-          }
-        },
-        lockOpts: {
-          lockerServer: {
-            host: 'localhost',
-            port: 3231,
-          }
-        },
-        messageBrokerOpts: {
-          messageBrokerServer: {
-            url: 'http://localhost:3380'
-          }
-        }
-      });
-      var config = {};
-      service._startWalletService = sinon.stub().callsArg(1);
-      service.start(function(err) {
-        err.message.should.equal('test');
-        done();
-      });
-    });
-  });
 });
